requirements.txt
cryptography==43.0.0
jupyter==1.0.0
marpyter==0.1.0
